Create a store for the binding spell to keep track of which entities are bound to each other.
The binding spell needs to know which entities are bound to each other in events. The current implementation of using tags is not working due to the events not passing the necessary tag information with the entities that are being tagged. In order to do this we need to hold state of the entities that are currently bound in a static class. Utility methods are exposed for interacting with the spell.

TODO: This change is only Java ready. There may be some Minecraft specific nuances that I am not fully aware of. Please review thuroughly and let me know what areas need more guard conditions to ensure only the right entities or other boundaries are not being crossed.